requirements.txt
numpy
tensorflow
flask
pillow

gunicorn==20.1.0